# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the public pool api from stratum.farm",
        "title": "1oop Pool API",
        "termsOfService": "https://stratum.farm/terms/",
        "contact": {
            "name": "1oop Support",
            "email": "pool@stratum.farm"
        },
        "version": "1.0"
    },
    "host": "152.228.229.130:3000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/pools/{pool_id}/miners/{miner_addr}/csv": {
            "get": {
                "description": "Download miner specific data as CSV",
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "CSV"
                ],
                "summary": "Download data as CSV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the pool",
                        "name": "pool_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Address of the miner",
                        "name": "miner_addr",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Specify the data type (stats, payouts, earnings)",
                        "name": "data",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start time (RFC3339 format)",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End time (RFC3339 format)",
                        "name": "end",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "utils.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        }
    }
}